requirements.txt
# --- AutoGen core & extensions -----------------
autogen-agentchat[gemini]~=0.4        # pulls core, google-genai, vertexai
autogen-ext[semantic-kernel-google]   # SK adapter for Gemini via SK
autogen-ext[openai]==0.5.7
semantic-kernel>=1.0.0

# --- Pipeline utilities ------------------------
chromadb>=0.4.24
pymupdf4llm        
markdownify>=0.11.6
markdown             # For Markdown to HTML conversion
tiktoken>=0.6.0

# --- Tool utilities ----------------------------
biopython
europe_pmc
semanticscholar
crossrefapi
unpywall
httpx[http2]
xmltodict 
pandas 
pydantic 
ratelimit
beautifulsoup4
readability-lxml
playwright
aiohttp              # For asynchronous HTTP requests

# --- Project hygiene ---------------------------
python-dotenv
pytest
pytest-mock
pytest-asyncio       # For testing asyncio code with pytest
respx                # For mocking httpx responses
rich                 # pretty CLI logs
prompt_toolkit       # advanced input with proper readline support
# (ruff, black, pre-commit go in dev-requirements.txt)

-e .
